Bass: Handle sync info request in all cases
While sink ask for sync info request when BASS is not synchronized
to the broadcaster, first synchronized it and then send PAST.

StopBigMonitoring when all receivers are active and no pending paused sinks.
Remove paused sink when receiver is active or requesting for sync info.

Do not monitor BIG if sink is active or requesting for sync info.

While resume receivers source synchronization, do not modify source,
if PA is synced or requested.

Bug: 372522289
Bug: 366294688
Bug: 363168099
Test: atest BassClientServiceTest BassClientStateMachineTest
Change-Id: If728689e8b7d1ce3e571a14df7c0893bc00950e2