Revert "HearingAid profile: rm impossible null handler"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89f4e09b-0795-4ceb-a0f2-b0b27f1db9aa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id1df67a49df865b8e43690cc4101fd86a4329e28
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com