Revert "HearingAid profile: rm impossible null handler"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/35578952-7aab-4a76-a7bd-4c37c98d358d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3ce218123445285155eb17dbd080cfbf2b4cfe09
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com