Revert "Update PDL for the CS HCI according to the released BT c..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d8613b4-8ca8-4c0b-8fc2-8bcfef445401). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e6ad201ad409aec84af293af79c712d7cab087b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com