Revert "Make connection_manager dumpsys more concise"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d9183418-ccd8-4fb0-8557-317b4bd712a0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375303134
Change-Id: Ie21ca24c51d970d1f88ec350746ee3e510db4d26
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com