Revert "Make it obvious that connection_manager is separate from..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d9183418-ccd8-4fb0-8557-317b4bd712a0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375303134
Change-Id: I44a4d9979a20058c5b3e2e0d2ec36f0d7286acdb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com