Revert "use_unified_connection_manager removal"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bbe7b1b2-bd0a-4b0f-87e2-6cb0b1973d89). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I539787592c770d50d44b7c84ee78ef46e3db6e6a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com