Revert "Remove un-necessary post"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/166308fc-7dba-4c98-b60d-1480a722af84). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375702391
Change-Id: Ia8b54a2a51942917313f7ab5b5edc6dba25d8e8f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com