Revert "Untangled RESTORE_USER_SETTING message"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/166308fc-7dba-4c98-b60d-1480a722af84). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375702391
Change-Id: I697350f55ce0dac3047cacdafb57492a6ce51815
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com