Revert "flags: Add adm_fix_disconnect_of_set_member"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f9d1451a-b615-402e-bce1-7a156046d31e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0d1827f23d02cb6fb6e41c273ef1355952ea288d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com