Revert "CS: Set tone_antenna_config_selection based on property"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ceb5a67e-6de1-48b1-b0fe-9cb481a7d2d5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0cd26eae7e9d770302f2c9dbccb3dd519a47cc8d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com