Return all attributes requested for GetElementAttributes
A previous patch enabled returning empty attributes for the
GetElementAttributes when all attributes were requested. As discussed
with partners, some carkits rely on empty attributes to display their
UI and the Bluetooth specification is not clear about what should be
done in the case an attribute is not known on the target side, so we
can add them empty as this won't have a negative impact anyway.

Bug: 346721681
Bug: 362896963
Flag: com.android.bluetooth.flags.get_all_element_attributes_empty
Test: atest BluetoothInstrumentationTests
Change-Id: I2cc55756e7afb134b144888a60eca2bd706058ac