Revert "flags: Add adm_fix_disconnect_of_set_member"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1e72d771-412f-40a8-ada8-36890074b0db). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie1e03a2fa213e415612189622f4d29dae10dac55
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com