Fix state machine stuck after disconnection
1. Connect on LE transport to dual mode device
2. Initiate pairing, control block will contain RPA as remote address
2. Obtain Identity Address, perform CTKD, generate Classic keys
3. Connect on Classic transport, start encryption
4. Encryption success, try reading the key size
5. Suddenly, connection times out, disconnection event is received

Current behavior:
Disconnection is not matched with the pairing device address.
We stay stuck in pairing state.

With fix:
We check if RPA from control block belongs to same device that
disconnected, and if so pass this event to state machine that would
clean up the control block.

Bug: 336242958
Bug: 369095269
Test: none
Flag: will add
Change-Id: I53bc0a4a8ac124dc20b4e9a1a30bb579653ef5a1