Revert "Remove obsolete GD style guide"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f24498b9-6050-4cb5-aa00-794f3455c0f5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 376042974
Change-Id: Ie791dd8fea629db9a2a7d8c68659b086da2c30dd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com