Rename AccountManager Account/Authentication Service and resources
This change updates the naming scheme of the account manager
authentication service classes and the paired resource strings. It makes
it clear that these items belong/relate to PBAP Client, as opposed to
"pbapsink" or nothing in particular at all. Docstrings are updated with
details around what the objects are for and how they work.

Bug: 365626536
Flag: EXEMPT, name/logging updates  only, no logic changes
Test: atest PbapClientAccountAuthenticatorTest.java
Test: atest PbapclientAccountAuthenticatorServiceTest.java
Test: m com.android.btservices
Change-Id: I513f7de71ba6826b7ea9ec11e4ea3256073482f3