Rename PbapClientVcardList to PbapPhonebook and update test coverage
This change migrates the VcardList object to be a more generic
"Phonebook" object, which can also own constants related to the
phonebook concept in PBAP, like the path names.

Additional test coverage was added for PbapPhonebook. This object owns
VCard parsing, so parse tests were moved to this object as well,
removing the need for a parser specific test.

Bug: 365626536
Flag: EXEMPT, mechanical refactor with no logic changes + test coverage
Test: atest com.android.bluetooth.pbapclient
Test: m com.android.btservices
Change-Id: I9a8896c6aaccf7e6f972c658625189d3efac5f5f