floss: mgmt: Wait until btadapterd completely stopped before restart
It's common to see btmanagerd is trying to restart btadapterd but
actually nothing happens, because the restart could be rejected by
upstart if btadapterd process is still alive (btmanagerd considers it as
stopped as soon as the pid file is removed, which is a bit earlier than
upstart).

Wait until btadapterd is completely stopped to avoid a redundant wait.

Also revise the NativeInvoker to match the new definition of the
ProcessManager API. For systemd and upstart invokers, comment is updated
and no behavior change.

Bug: 374017641
Tag: #floss
Test: mmm packages/modules/Bluetooth
Test: modprobe -r btusb; modprobe btusb  # btmanagerd waits for
      btadapterd to completely stop, and then restarts it.
Flag: EXEMPT, Floss-only change
Change-Id: I99192d3db64cd7ae29726425ae1c4315ef738514