osi: Use CLOCK_MONOTONIC on platforms that can't wake on BT
Some devices (such as TVs) are not wired up to be
woken by the controller. These devices might run
into unforeseen issues going to sleep with timers
armed. This commit adds a sysprop to specify that
the platform is wired up like that, and makes all
alarms on those platforms monotonic, so they do
not count up during suspend.

Bug: 343826774
Test: m -j128 com.android.btservices & QA testing
Change-Id: Ie720ea3c20cba317eb3536c49604c9c8b9c2d80f