Revert "CS: Set tone_antenna_config_selection based on property"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/00aa464b-92d2-480e-86d8-159f2f44fa46). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib22b6e3b4be264ae62e14e847829cca12ce74a56
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com