Revert "Add flag: smp_state_machine_stuck_after_disconnection_fix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2ef2f9ef-d0d0-46db-b584-53ba1cf0bbfa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib3f4d2c7d585cee50d291e9501d933bf7fca46a6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com