Revert "CS: Retry when config creation fails"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5a8e0d77-9f9a-4b37-a6ac-074bf8423b68). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I51cfc5269751861153669835aab22dffc9523c24
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com