Revert "Add flag: smp_state_machine_stuck_after_disconnection_fix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5a276492-1d86-4be4-b691-f6623e755015). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3aa0b9ef447a63cb6555229e6088b6ff89c1235a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com