Storage Refactor: Introduce a new PbapClientStateMachine
This new version of the PbapClientStateMachine is now responsible for
collecting downloaded contacts and storing them in the new storage
abstraction. The download process has been updated to account for this,
where the state machine now owns batching decisions, OBEX requests,
storage (of course) and clean up when the device disconnects.

With this new state machine, PbapClientService must now delegate,
based on flagging, to either the old implementation, or this new one.

Unit test coverage has been provided for all new code.

Bug: 376461939
Bug: 365626536
Test: atest com.android.bluetooth.pbabpclient
Change-Id: I1f56194d6ff15359d8b6f4c6c18623fc37022239