Revert "Add avrcp_16_default flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/664d2227-0321-4a6e-9d15-b13c60a5cc3b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idf6f1f28dc376ed318fd213911ad260c00bde0f9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com