Revert "SystemServer: reduce callback visibility"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d2b3c976-b007-469a-9dc7-8515469f1ee6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I083805969484fa3af23b97db463447c31f2bc2ab
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com