Revert "Fix empty data mask detection"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a234213-dfd8-44cd-b863-c846ca5c3a39). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib2e60af6f96ff49dc04ef08bddcefbdf1b68d210
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com