Revert "Fix empty data mask detection"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8c95a99d-5cf9-491b-b64f-134f93b0f066). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia5aec0fbc8bcd11ad554e16f33d56a6dab1cbc3e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com