Revert "leaudio: Do not update local metadata if not needed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/249a2679-dae7-40f4-89c9-3eeb0d9058e4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I56ea4afc356945cded3c730b5f930bac941101bb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com