Revert "leaudio: Do not update local metadata if not needed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1e70e7d5-2afe-4557-9367-c97a5fcc1b84). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4a80a80ab894c54ff33f3e05d9f447b476fdcf99
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com