Revert "leaudio: Do not update local metadata if not needed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/847e56bb-0ec3-4181-a266-499756733748). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6c238fb514e0eb007066ffa116a15f6231f1c142
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com