Revert "leaudio: Do not update local metadata if not needed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b154c9e9-5a67-4974-a9c2-71ce53f739c4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I376c04a6accd2441feea764357ff6a123f164960
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com