Revert "Add socket settings interface"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bed69dc7-f4d7-4245-9e04-4b128f463624). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1d0abb13456ddc3fdc01f1dd79f65d10e330fb93
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com