Revert "leaudio: Do not update local metadata if not needed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/150a04b1-9b0d-42fd-ba25-47a7c8cb81b2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iab8656bc248e660bd7db1375d421f866b9f0dc7e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com