Revert "Handle active device when get the audio policy value fro..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/19d7aae0-0b08-4023-ba3f-af80323f131d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibe652ccad2532edb94b1ff5eff857fbd14d75297
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com