Revert "leaudio: Non functional change"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b8b9227-5968-41aa-b4ea-5100aa6e1216). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9856ccb473128d06037c262e737a4b5001ef6d8a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com