Revert "CsipSetCoordinatorService: Fix possible null ptr dereference"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b8b9227-5968-41aa-b4ea-5100aa6e1216). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iade20f8a98c768ea861e3c0d78dfb4aa89e3d2c9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com