Revert "btm_io_capabilities_rsp: remove unnecesary null check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/810f5435-b3e7-4593-ae66-e15fbd20949d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I36bafffc09d8952b5a1e9e2c15a4886ceb717060
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com