Revert "btm_io_capabilities_rsp: remove unnecesary null check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/baf8a76e-0b26-43f5-aeb8-77b9ff271e4f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie797ff01d0c27d00b8ac5e1a30c7bc5397af7d8c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com