Check rcb is released
When BT is disabled, a race condition can
occur between the release of a registration
control block (RCB) and an attempt to access it.
This can lead to accessing deallocated memory.

This change resolves the issue by checking if
the RCB is still valid before accessing it.

Bug: 377818132
Flag: Exempt, strict checking
Test: atest BumbleBluetoothTests
Change-Id: I63cc4312d544ac6b4dee4b155b18a7aa1f44f794