Revert "Checking if the channel sounding is supported by controller."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9cb22ae-536b-41a5-a52f-1e29db640289). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61761cd35b910a51589df415d1f652eecd553d92
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com