Contacts Caching: Implement PBAP Contacts Caching
This change allows the contacts downloaded over PBAP to be persisted
across device connections, in order to make contacts available quicker
on subsequent connections.

It does this by using the specification defined database identifier and
version counters to determine if the locally saved contacts are the same
as the remote contacts. If they are the same, the download is skipped
and the contacts are instead made available. Any discrepancies in
metadata between our local device and the remote device go in favor of
the remote device.

Contacts are to be made entirely inaccessible while a device isn't
connected. They are deleted when the connection policy changes from
allowed, or when a device unbonds. Contacts that live on the device
without access or update for 7 days will be removed.

Bug: 376461947
Bug: 379107040
Test: atest com.android.bluetooth.pbapclient
Change-Id: I859ed5af00e1d3f796de161f9b4e446093291298