Revert "cpplint: fix and enable whitespace/ending_newline"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/405f1db3-6b6d-40d2-a89a-417697d79c4a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7c3b308c530cdd3eb00cd466789d7ab452d802a6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com