Revert "Add new metadata APIs get/set for is_microphone_for_call..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/61d62b09-1f21-44b1-96da-d684a22d39d3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5a7c1d0e15aff6c5c277eb516c5412b75563fc67
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com