Revert "Fix some VisibleForTesting usage problem in Metadata"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/61d62b09-1f21-44b1-96da-d684a22d39d3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If5a6ee0747624d4a6697d43e420d328720bdfdde
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com