Revert "Test: Update tests related to bond check from Profiles"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c3acd8f9-3ed4-484d-bfe3-8289476d6155). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379274275
Change-Id: I980c4f1eeb90e49327a7d3426cde6f0577bb81e8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com