Revert "AICS: pre-API definition and intdef"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05d8b351-c799-427c-ae3e-a339b2c6cdae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I136051f7441850c6e887a6b2412b5d8ececb1082
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com