Revert "AICS: description set/get/onChange implementation"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05d8b351-c799-427c-ae3e-a339b2c6cdae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia0ec71c74acce5e9e8d5f0e9f693e056750c95de
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com