Revert "AICS: cleanup and preliminary changes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05d8b351-c799-427c-ae3e-a339b2c6cdae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I53ef903b63244c538901def31c8e82910e86fd0a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com