Revert "BluetoothMetrics: Log LE ACL Completion Event"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/20f7eb66-ec78-4a12-b0bf-3eb7f7478c90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379377240
Change-Id: I0375a03c41dfef9613a691226ae547dcd18f961d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com