Revert "RfcommTest: ConnectToServer and Disconnect"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3ccacc30-a30b-44c6-899e-1f7765daf89d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5dba745a17eb12d61cf325febabbc86827b97ff1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com