Revert "RfcommTest: ConnectToServer and Disconnect"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/de3258bb-b60a-464b-9ba4-2600752d366c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I99714a865c2335ddef4f4b33e0743e5ded176145
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com