Revert "BluetoothMetrics: Log LE ACL Completion Event"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/439a94e5-7b4c-4291-aecc-c27478fdc03a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I03b9da61357bd1f42684d9f4c2a6056a0515c2ed
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com