Revert "CS: Retry when config creation fails"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2283c9fd-5e46-4f8c-8179-89c3e03a3a5b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8cc91cef2d6bf2e59685e6a408233439dc755432
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com