Revert "CS: Retry when config creation fails"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a394dd87-d289-4b8d-b1b8-c2d5fe78e0ed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia171ad9cd400a6d997b8ab3803b2e788f428f685
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com