Revert "Factorize system/packet/avrcp/tests/fuzzers/Android.bp"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d6887c7-2796-4fb4-bf0d-78e491c66745). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I845fc4a23e72717d1e1a8b136f376e39838c47bd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com