Revert "flags: Add rpa_offload_to_bt_controller"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/87bbc7da-4440-4e2e-a8cd-99e5c60dc9cf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifc1e54f94710b259e6f72eab6f9648c3deecb987
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com