Revert "RfcommTest: ConnectToServer and Disconnect"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8e64428-9263-4121-aea2-9681b0355ea9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib6e36f12a24ca5f43c4f3bada7343e06fdaf1f5d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com