hh: Reconnect HoGP in the background on local initiated disconnection
When disconnecting HoGP locally, we don't rearm the GATT client. This
potentially causes the peer to be not added to the accept list, or
causes incoming input notification to be ignored on reconnection.

This CL sets a background connection when a HoGP device is
disconnected, unless it is not allowed by the host.

Bug: 378011758
Bug: 378162528
Test: m -j
Test: Manually, by disconnecting the device from the UI or shell
Flag: com.android.bluetooth.flags.hogp_reconnection

Change-Id: I69487a336c2c4f2bc0a35648638c2f5ee65fd601