Revert "Remove public visibility on BluetoothPacketSources"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c511c949-ec4d-4855-973d-a76c4e3bdd48). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380202258
Change-Id: I9ff30a967c0e4bcaf81277cd6f17bedb5ef82690
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com