Revert "Don't leak internal flag to CTS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e0a19387-2436-4aa7-87c2-35084ed82d75). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380709792
Change-Id: I853f95c5b054332770d62c801c855b4aeecfa7a1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com