Revert "Don't leak internal flag to CTS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0e3dde67-6be4-4c34-97ed-6ecc51109b16). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380709791
Change-Id: I33dfdb31094dc58871fee04c2cc5058454cc061c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com