Revert "Don't leak internal flag to CTS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1fa06a0f-2574-47ee-8001-c99b3c2b3db2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380709747
Change-Id: I48810ef9489321b40222c258f397336f9b13726f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com