Revert "VolumeControl: Fix canceling opportunistic connect"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/02abce7c-cab9-405a-bd5d-644bfd356080). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1ea1d85531a897a52fe285b6bc5acd99f5cc4e38
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com