Revert "Avoid processing l2cap events when stack being shutdown"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6b3e775d-6a8f-4804-802a-43575ab29c71). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6f7fcee5f24607ebeb0499f90446f4e4fdd20391
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com