ScanManager: fix flaky test
The ScanManagerTest is flaky since 2022 when it was created. Because of
race condition in how the test was written.

This CL is using the test looper and only dispatch message when
expected. It also inject a time provider to advance the time by the same
amount

Bug: 330956986
Fix: 330956986
Bug: 380539632
Bug: 380510617
Bug: 380510672
Bug: 380536618
Bug: 380510672
Bug: 380510672
Bug: 380536573
Bug: 378542424
Bug: 346028665
Bug: 347474468
Test: atest ScanManagerTest --rerun-until-failure 3000
Flag: TEST_ONLY || The change in app are no-op
Change-Id: I7c7d397453328defa50fe623454f078f36253f22