Revert "Remove gatt_cleanup_restricted_handles flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d3fbb021-d136-4565-ae69-c61e008abd38). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic3b3738c573f27a25169edc510db58d0b38cc83f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com