Revert "Avoid processing l2cap events when stack being shutdown"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0e7b25b1-52c9-4964-87fa-6107c03b3e9f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2e7a8bf124e500fb41d26a50d16784c61870060d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com