Cancelling pending connections to already added HID device
Initiated connection to the already added HID device puts it in the pending connections list. If application requests disconnection before connection is completed, the device control block state is updated but the device is not removed from the pending connections list.
This allows the remote device to be able to reconnect even when the connection policy forbids it.

Change-Id: I093bd17fce02989048bfec99e5986dc91e5acb04
Test: atest BumbleBluetoothTest:HidHostTest
Flag: com.android.bluetooth.flags.pending_hid_connection_cancellation
Bug: 380477704
Bug: 381135214