Revert "Don't leak internal flag to CTS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e69793d-e9af-4584-b7f6-40839c51e8ad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 381151048
Change-Id: Idf414d60de119d4d9ed93147a4af7af963e3bb64
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com