Revert "Writes procedure data for HAL V2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6bb710ba-4f53-43ef-81b9-c4f1e19d5c50). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 381337331
Change-Id: Ie04f99014dd51d521f5e46afac724f3f14793301
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com