Revert "topshim: a2dp: Avoid leaking the reference of a stack object"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4f4a696-4d42-4947-b9ca-ca8b6bd61630). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2e2c31119b86b57571941192607778713e6a037a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com