Revert "topshim: a2dp: Avoid leaking the reference of a stack object"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5c55488a-d3aa-4d80-83ed-63b003ef902d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0c43d9e4b02aa621d7a687e3e0f7e0e413e80b0b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com