Revert "topshim: a2dp: Avoid leaking the reference of a stack object"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/973a897c-4438-4634-81fb-0fa2db1041d5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id0f408be1058a0a2a58b623e024613c5773a1986
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com