Revert "topshim: a2dp: Avoid leaking the reference of a stack object"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f4fd6e1c-eabe-41d8-8f01-565c1455140f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If569f58291ae752d9941963c469c0bb774cb186d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com