Revert "Writes procedure data for HAL V2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb4c74ec-655b-4e00-884f-5c0442fbba9c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1837184e56fd8cba8f7339b55513d333474253e6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com