Revert "Flag to properly cancel the pending connections to the a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d398502e-de46-4f62-83de-c95093e9a059). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I16f88a97e89eb1c474975f6ae5cf5faf91ae6b2c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com