Revert "VCS: Add wrappers for getGroupId and getGroupDevices"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d5f5fe84-fffe-42d6-b2bc-817668abe836). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383404401
Change-Id: I257a73c07cff28f86c0b2f65390d875687672798
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com