Revert "VCS: Add wrappers for getGroupId and getGroupDevices"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7b2eb995-f8a9-481d-857a-75b3d2669b32). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383403712
Change-Id: I279e1d43ba073e910486cc0ce6a0bb957493f434
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com