Revert "Revert "Make SetBrowsedPlayer return current path and co..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/18e50d1f-3d10-4a12-87ce-4c35e3e3a2b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I553aa7a2bc32725ba629e365af637713884baaa6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com