Configure scan on resume if necessary
New scan parameters was not being applied if scan parameters
were updated while the scanning manager was paused.
This was due to `configure_scan()` not being called in `onResume()`.

This change ensures the scan is correctly configured when the
scanning manager resumes. To prevent redundant calls to
configure_scan whenever it's resumed, do it only when
SetScanParameters is called.

Bug: 381010390
Test: atest bluetooth_test_gd_unit
Change-Id: Ia0e93d2a6f8416cbd4fb0678d57c8deee9f1229b