Revert "leaudio: Improve cleanup function"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0373afbd-ea08-4727-b781-d9e27a0903f5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iae0716a9f92d54bc04bd42f041289ad3471a9756
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com