Don't use config object after std::move
Problem: Devices marked as restricted were not properly getting removed
after existing restricted move. This was because we were trying to call
the removal function on an object after calling std::move() on it. This
caused the function not to operate on the same effective object and the
devices weren't properly found and removed from the cache, nor was the
file updated.

Solution: Use the reference of the config now owned by pimpl_ so we act
on the proper set of objects.

Bug: 357640474
Flag: EXEMPT, simple bug fix
Test: m com.android.btservices
Change-Id: Iebfe73efa6eca7302cc7866196cfacb13185ad68
(cherry picked from commit 39c4fa71a5ae2bafeb8700c4a3f2e11858669f78)