Revert "Revert "Make SetBrowsedPlayer return current path and co..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4728641f-a6c9-4e34-b351-0fc7d9e2b1a2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I55e05f4eb137fceb7f04bb94076f48e9d50bf033
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com