Revert "One second delay introduced to avoid setting app interfe..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1f4ea80-9a71-43cd-8188-24d9c29235ee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I77ce4101aed8556cee29f129d0ed2f4a75fe2442
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com