Revert "Update the advertising_event_properties to match the def..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ed925c3-4d44-4768-8def-030b9710f97b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I13681c85480e72442e38b9c91e419688782e5ddd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com