Revert "Comment cleanup: btif_sock_l2cap.cc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2cdfaf9b-078b-4410-b9c3-71b6d3ac8d24). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8ac3139ec57beeb6995a703a942299e9d5a9ef26
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com