Revert "RAS: Prevent null pointer dereference while the service ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2cdfaf9b-078b-4410-b9c3-71b6d3ac8d24). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I454338f984dc909eac5e42f6312cee08a925e77e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com