Revert "Comment cleanup: btif_sock_l2cap.cc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2903a3b2-ce14-4f6c-aff2-8e5e504159a0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id1e0389d3f4c4e7b33e591bc1310043ef695b7b8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com