Revert "RAS: Prevent null pointer dereference while the service ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2903a3b2-ce14-4f6c-aff2-8e5e504159a0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I62a14862f1a2969890abe042335e5f7ea34e364f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com