Revert "bluetooth: configure wake by BT according to device state"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd918c57-72ec-4ac5-ab6a-5d9915488f5d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I91f03e93eb6b7f105b123a5208990c2148820b84
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com