Revert "Comment cleanup: btif_sock_l2cap.cc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd918c57-72ec-4ac5-ab6a-5d9915488f5d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I15b0524d653e3b1987d2a7eebc3ab4f9acf175f6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com