broadcaster: Add missing fallback default update handlers
This change modifies two cases of handling update of fallback Unicast to
Broadcast group Id. For first case the fallback group should not be
updated when group is changing state to inactive. Second case is for
group from which active device (device which is part of active group)
is disconnecting.

Bug: 384847958
Bug: 375422795
Flag: leaudio_broadcast_primary_group_selection
Test: atest LeAudioBroadcastServiceTest
Change-Id: Ic18f32caabd6603a6762b9abcdfcfb2b18a6d510