Revert "Change quiche path to Rust crate monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b7a152fe-b3b8-431c-b80a-231bac214df6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I50fffd925dd56ad4d37e6d5aaecb8e0642f646eb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com