Revert "Errorprone: Fix & enforce RedundantControlFlow"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/548827ee-81d7-4fdc-8751-bce9691b175e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib223b93299f4cee62b18a3e607e95780fcd63f63
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com