Revert "Comment cleanup: btif_sock_l2cap.cc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/87ae85e1-f28e-4e9f-8231-da908a191f3f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id85b0f948940548af726bf139c5b6e733c9ccbf3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com