Revert "RAS: Prevent null pointer dereference while the service ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/87ae85e1-f28e-4e9f-8231-da908a191f3f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2446568c0565056b31147fdb2f36cd3d33ad8a9a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com