Revert "Flag to retain identity address on BT restart"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b699550d-7625-46f6-b393-c8d40ced9361). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4b459b4354f15fa5f90cf2adbd5ef243005ddb77
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com