Revert "Errorprone: Fix & enforce DuplicateBranches"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b699550d-7625-46f6-b393-c8d40ced9361). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I10142ba4ac53592201e2fec9f938ed5e4876ae9c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com