Revert "RAS: Prevent null pointer dereference while the service ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1e3021d4-82f1-4bb7-86e9-85c19aea612c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1be29125d1ff3ea0ec728465f24c6922b2aead9b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com