Revert "bluetooth: configure wake by BT according to device state"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bdde8b0a-47c0-417d-bb00-4b11abf1b651). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4713b6d8448e4c8fea0ea052bdbcff2708ebcb8c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com