Revert "flags: avdt_handle_signaling_on_peer_failure"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f75eef53-04a2-41c0-b9de-85691cb1310f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I25cd73c4ced39044c6257fd273c66d2fe4901616
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com