Revert "RAS: Prevent null pointer dereference while the service ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/61387436-6e49-4be6-8661-fbbd0691bc45). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8cdca428ebe1e50fc1c0c4a531b5bb1fc898c0d0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com