Revert "Drop messages with handles we didn't ask for"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/41cf4a24-1406-4e1e-87d4-fb083c2fe126). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8de4e2a36cacd59096840111a80944cd5d752d2d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com