Revert "RAS: Prevent null pointer dereference while the service ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a73b9ff-9a79-4385-8be2-0ae4c6b2318e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b6d9f93936944b9e0c42c0d229c203a1770cdd9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com