Revert "Remove gatt_rediscover_on_canceled flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f4756973-ca56-49d8-b908-b2fbb7e237b2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 384619750
Change-Id: Ic320e51cd24a5249d3e66a8e6e19b0598a4ef3dd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com