Revert "Remove blueberry facade proto definitions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/524497a1-832c-4589-852f-f2f18e71556e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 384933064
Change-Id: Ic8d7709627ede2bd091e3cf1193189d386f48152
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com