Revert "Remove blueberry facade proto definitions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e088a884-8b5c-4422-9e08-90bbfd5000e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 384934056
Change-Id: Ida2f4944900632494ed87e64daeb3e147a2e9002
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com