Revert "Remove the rust_binary bt_topshim_facade"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e088a884-8b5c-4422-9e08-90bbfd5000e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 384934056
Change-Id: Ie35e1767ce5af4f06c3472047fa08a7985f9907d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com