Revert "remove le_scan_use_uid_for_importance flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b4408c28-a8c7-4c7b-bdae-4d8419e5f03f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385168972
Change-Id: I31878a659b71e54fbd7c72917b17552405fffe09
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com