Revert "Comment cleanup: btif_sock_l2cap.cc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7cb56198-cb5a-4671-82f5-55d59f2c6744). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4aec2df904d5963977030843b5cb5141761c61f8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com