Revert "RAS: Prevent null pointer dereference while the service ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7cb56198-cb5a-4671-82f5-55d59f2c6744). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I65494f5e62f3efd7ff05f2064728a45a65521108
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com