Revert "Batch call log insertion operations"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/93e558aa-f56c-476c-8be3-b99dce7d243f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5a3a00464d1ffdb6a09d2dad750d168f7fcf0358
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com