Revert "Comment cleanup: btif_sock_l2cap.cc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a7b7b205-62c9-47af-a326-87164ae7cf63). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7abe6c870fabc6128156ca0f4a7335343159d933
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com