leaudio: Inactivate device on autonomous stream release
If remote device wants to stop the stream then either GTBS or GMCS should
be used to terminate the call or stop media stream.
If remote device just drop the LeAUdio stream, this is hard for the
Initiator to find out the intetions of the remote device.
From now on, when this happens, device will be considered as device not
willing to be used for streaming and removed from ACTIVE devices.

Bug: 314091632
Flag: Exempt, regressions checked with unit tests
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test
Change-Id: I9a2b5ba06a907e410a3498215eaf16cf7021af3a