LeAudioService: When connectGatt coexistance with LeAudio device
When application is connecting using connectGatt to device which
supports LeAudio, that device will not get connected to Audio Framework
automatically.
Instead it will stay in the background waiting for user action to make
it Active or Remote device starts Targeted Annoucements.

If device will not became Active and disconnectGatt will happen, then
device will be disconnected as LeAudio service will not keep the
connection.

This is implemented in order to allow 3rd party applications to
connect to LeAudio device without intention for Audio streaming.

Bug: 388179850
Test: atest LeAudioServiceTest GattServiceTest AdapterServiceTest
Flag:
Change-Id: I2a1f4f0a5fcf28f8fc434f19e54870166463d784