Revert "Fix incorrect previous active device"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e488605-a21c-4424-b6a6-b9be9145d04a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idbda7124786c07d730297a70a7eab91e9e02fe77
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com