Revert "Remove Sockets 24Q3 flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2297154a-becf-4268-8fa7-62121d0515c9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1875791b8e860a2cdc592e02540bb24c68c47789
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com