Revert "system/gd: Merge StackManager into shim::Stack"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2297154a-becf-4268-8fa7-62121d0515c9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idf3d2fd9800b79d73dd54cbb20bfe6e5e53a8b1b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com