Revert "Errorprone: Fix & enforce DuplicateBranches"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f50869d8-d67d-443f-ae06-caf4ff335fc8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id99e96fec60e538cf9250361cc0120b8ae679bcc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com