Revert "VolumeControl: Fix multi-read request"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ae3e4842-8b15-461b-a278-4d8d13dd82e3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I053aa39c178db02749f9ddc65453ed055139514f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com