Revert "pbapclient:: Check SDP record callback status for validity"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/750af4c3-bbbd-4cd3-8e6f-7fbc2da1c2cd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9bbd2be17877c45b9c4b3650b6f9effb4526a05b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com