Revert "Update the doc of getChannelSoundingSupportedSecurityLevels"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/032f8e76-0c0e-48ec-8cc4-9f5dab31abfb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I00eb49826acd32105ca5f1255846e44d3d416562
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com