Revert "24Q4: Removal of override_context_to_specify_device_id"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/923eec2e-226a-4fe7-82b1-471d532a0bb1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I10505edd24033d661b23ceec3cd4b47f00d42ef6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com