Revert "24Q4: Removal of respect_ble_scan_setting"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/923eec2e-226a-4fe7-82b1-471d532a0bb1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I722fcc975a9440f7b3f5ab9547018412d5c0fccf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com